Add remote branch not fully merged protection

+++ gmd/Cui/Common/RepoCommands.cs
@@ -697,15 +697,21 @@ class RepoCommands : IRepoCommands
         {
             if (!Try(out var e, await server.DeleteLocalBranchAsync(localBranch.Name, rsp.IsForce, repoPath)))
             {
-                return R.Error($"Failed to delete local branch {branch.Name}", e);
+                return R.Error($"Failed to delete local branch {localBranch.Name}", e);
             }
         }
 
         if (rsp.IsRemote && remoteBranch != null)
         {
+            var tip = repo.Commit(remoteBranch.TipId);
+            if (!tip.ChildIds.Any() && !rsp.IsForce)
+            {
+                return R.Error($"Remote branch {remoteBranch.Name}\nnot fully merged, use force option to delete.");
+            }
+
             if (!Try(out var e, await server.DeleteRemoteBranchAsync(remoteBranch.Name, repoPath)))
             {
-                return R.Error($"Failed to delete remote branch {branch.Name}", e);
+                return R.Error($"Failed to delete remote branch {remoteBranch.Name}", e);
             }
         }
